feat: move logic to common file
